+++ src/extensions/multiselect/multilist.js
@@ -3,9 +3,20 @@
  * @fileOverview
  * @ignore
  */
-define('multilist', function(require){
+define('bui/extensions/multiselect/multilist', ['bui/common', 'bui/list'], function(require){
   var BUI = require('bui/common'),
-    Component = BUI.Component;
+    Component = BUI.Component,
+    List = require('bui/list');
+
+  //设置Controller的属性
+  function setControllerAttr(control, key, value) {
+    if (BUI.isFunction(control.set)) {
+      control.set(key, value);
+    }
+    else {
+      control[key] = value;
+    }
+  }
 
   var PREFIX = BUI.prefix,
     CLS_SOURCE = PREFIX + 'multilist-source',
@@ -45,7 +56,7 @@ define('multilist', function(require){
       if(config.isController){
         return config;
       }
-      var multipleSelect = config.multipleSelect,
+      var multipleSelect = config.multipleSelect || this.get('multipleSelect'),
         //如果已经传了xclass则优先使用xclass
         xclass = config.xclass || multipleSelect ? 'listbox' : 'simple-list';
       config.xclass = xclass;
@@ -93,19 +104,72 @@ define('multilist', function(require){
           source.addItem(item);
         }
       });
+    },
+    //设置store时直接过滤选项
+    _uiSetStore: function(store){
+      var _self = this,
+        target = _self.get('target'),
+        idField = target.get('idField');
+      store.filter(function(item){
+        var items = target.getItems(),
+          flag = true;
+        BUI.each(items, function(i){
+          if(i[idField] === item[idField]){
+            flag = false;
+            return false;
+          }
+        });
+        return flag;
+      })
     }
   }, {
     ATTRS: {
       source:{
+        value: {
+          elCls:'bui-select-list'
+        },
+        shared: false
       },
       target: {
+        value: {
+          elCls:'bui-select-list'
+        },
+        shared: false
+      },
+      store: {
+        setter: function(v){
+          setControllerAttr(this.get('source'), 'store', v);
+          return v;
+        }
+      },
+      items: {
+        setter: function(v){
+          var store = new BUI.Data.Store({
+            data: v
+          });
+          this.set('store', store);
+          return v;
+        }
+      },
+      url: {
+        setter: function(v){
+          if(!v){
+            return;
+          }
+          var store = new BUI.Data.Store({
+            url: v,
+            autoLoad: true
+          });
+          this.set('store', store);
+          return v;
+        }
       },
       tpl: {
         value: '<div class="row">' +
                 '<div class="span5 ' + CLS_SOURCE + '"></div>' + 
-                '<div class="span2 bui-multilist-action">' + 
-                  '<button  class="button button-small ' + CLS_BUTTON_RIGHT + '" type="button">>></button>' + 
-                  '<button  class="button button-small ' + CLS_BUTTON_LEFT + '" type="button"><<</button>' + 
+                '<div class="span2 bui-multilist-action centered">' + 
+                '<p><button  class="button button-small ' + CLS_BUTTON_RIGHT + '" type="button">&gt;&gt;</button></p>' + 
+                '<p><button  class="button button-small ' + CLS_BUTTON_LEFT + '" type="button">&lt;&lt;</button></p>' + 
                 '</div>' + 
                 '<div class="span5 ' + CLS_TARGET + '"></div>' + 
               '</div>'
