--- src/mulit/multilist.html
@@ -1,55 +1,0 @@
-<!doctype html>
-<html lang="zh">
-<head>
-  <meta charset="UTF-8">
-  <title>SuggestList demo</title>
-
-  <link href="http://g.tbcdn.cn/fi/bui/css/dpl-min.css" rel="stylesheet">
-  <link href="http://g.tbcdn.cn/fi/bui/css/bui-min.css" rel="stylesheet">
-  <script src="http://g.tbcdn.cn/fi/bui/jquery-1.8.1.min.js"></script>
-  <script src="http://g.tbcdn.cn/fi/bui/bui-min.js"></script>
-  <script src="suggestlist.js"></script>
-  <script src="multilist.js"></script>
-  <script src="multiselect.js"></script>
-  <link href="multiselect.css" rel="stylesheet">
-</head>
-<body>
-
-  <div id="J_MultiSelect">
-  </div>
-
-  <script type="text/javascript">
-
-    var items1 = [
-        {text:'选项1',value:'a'},
-        {text:'选项2',value:'b'},
-        {text:'选项3',value:'c'},
-        {text:"数字值",value:3}
-      ],
-        items2 = [
-        {text:'选项1',value:'a'},
-        {text:"数字值",value:3}
-      ];
-
-    BUI.use(['suggestlist','multilist'], function(SuggestList, MultiList){
-      var store = new BUI.Data.Store({
-        //data: items1
-        url: 'data.php',
-        autoLoad: true
-      });
-      var suggestlist = new SuggestList({
-        store: store,
-        elCls: 'bui-select-list',
-        multipleSelect: true
-      });
-
-      var multiselect = new MultiList({
-        render: '#J_MultiSelect',
-        store: store,
-        elCls: 'bui-select-list'
-      }).render();
-    });
-   
-  </script>
-</body>
-</html>
