--- src/mulit/multiselect.html
@@ -1,53 +1,0 @@
-<!doctype html>
-<html lang="zh">
-<head>
-  <meta charset="UTF-8">
-  <title>SuggestList demo</title>
-
-  <link href="http://g.tbcdn.cn/fi/bui/css/dpl-min.css" rel="stylesheet">
-  <link href="http://g.tbcdn.cn/fi/bui/css/bui-min.css" rel="stylesheet">
-  <script src="http://g.tbcdn.cn/fi/bui/jquery-1.8.1.min.js"></script>
-  <script src="http://g.tbcdn.cn/fi/bui/bui-min.js"></script>
-  <script src="suggestlist.js"></script>
-  <script src="multilist.js"></script>
-  <script src="multiselect.js"></script>
-  <link href="multiselect.css" rel="stylesheet">
-</head>
-<body>
-
-  <div id="J_MultiSelect">
-    <input type="hidden" id="J_SelectVal">
-  </div>
-
-  <div id="J_MultiSelect1">
-    <input type="hidden" id="J_SelectVal1">
-  </div>
-
-  <script type="text/javascript">
-
-    var items1 = [
-        {text:'选项1',value:'a'},
-        {text:'选项2',value:'b'},
-        {text:'选项3',value:'c'},
-        {text:"数字值",value:3}
-      ],
-      items2 = [];
-
-    BUI.use(['suggestlist','multilist', 'multiselect'], function(SuggestList, MultiList, MultiSelect){
-      var store = new BUI.Data.Store({
-        url: 'data.php',
-        autoLoad: true
-      });
-
-      var multiselect = new MultiSelect({
-        render: '#J_MultiSelect',
-        valueField: '#J_SelectVal',
-        url: '',
-        source: items1,
-        searchTpl: ''
-      }).render();
-    });
-   
-  </script>
-</body>
-</html>
