--- src/mulit/suggestlist.html
@@ -1,76 +1,0 @@
-<!doctype html>
-<html lang="zh">
-<head>
-  <meta charset="UTF-8">
-  <title>SuggestList demo</title>
-
-  <link href="http://g.tbcdn.cn/fi/bui/css/dpl-min.css" rel="stylesheet">
-  <link href="http://g.tbcdn.cn/fi/bui/css/bui-min.css" rel="stylesheet">
-  <script src="http://g.tbcdn.cn/fi/bui/jquery-1.8.1.min.js"></script>
-  <script src="http://g.tbcdn.cn/fi/bui/bui-min.js"></script>
-  <script src="search.js"></script>
-  <script src="multilist.js"></script>
-  <script src="multiselect.js"></script>
-  <link href="multiselect.css" rel="stylesheet">
-</head>
-<body>
-
-  <div class="row">
-    <div class="span5">
-      <div id="J_SearchList">
-      </div>
-      <div id="J_MultiList">
-      </div>
-
-      <div id="J_MultiSelect">
-          <input type="hidden" id="J_Val" value="b">
-      </div>
-    </div>
-  </div>
-
-  <script type="text/javascript">
-
-    var items1 = [
-        {text:'选项1',value:'a'},
-        {text:'选项2',value:'b'},
-        {text:'选项3',value:'c'},
-        {text:"数字值",value:3}
-      ],
-        items2 = [
-        {text:'选项1',value:'a'},
-        {text:"数字值",value:3}
-      ];
-
-    BUI.use(['bui/list', 'search', 'bui/data', 'multilist', 'multiselect'], function(List, Search, Data, Multilist, MultiSelect){
-      var store = new BUI.Data.Store({
-        url: 'data.php',
-        autoLoad: true
-      });
-      var list = new List.SimpleList({
-        elCls:'bui-select-list',
-        render : '#J_SearchList',
-        store: store,
-        plugins: [Search]
-      }).render();
-
-      var multi = new Multilist({
-        render: '#J_MultiList',
-        source: {
-          elCls:'bui-select-list',
-          store: store
-        }
-      }).render();
-
-      var select = new MultiSelect({
-        render: '#J_MultiSelect',
-        title: '标题',
-        //items: items1,
-        url: 'data.php',
-        valueField: '#J_Val',
-        multipleSelect: true
-      }).render();
-    })
-   
-  </script>
-</body>
-</html>
